test-requirements.txt
# bugs
-e git://github.com/KeepSafe/aiohttp.git#egg=aiohttp

# coverage
coverage>=4.3.4

# flake8
flake8>=3.2.1
pep8-naming>=0.4.1
flake8-import-order>=0.8

# pytest
pytest>=3.0.5
pyfakefs>=3.1
pytest-asyncio>=0.5.0
pytest-aiohttp>=0.1.3
asynctest>=0.10.0
pytest-aiofiles>=0.1.0
